requirements.txt
flask
python-binance
requests
python-dotenv
